refactor: ENUM 구조 변경에 따른 변경
